rename cassandra.py, was causing namespace collisions with the cassandra driver
